Update is_group_name to compare lowercase names
Ordering the inbox from oldest to newest would break if you clicked on a notification for the PA inbox, because it'd link to `messages/group/post_approval` instead of `messages/group/Post_Approval`. post_approval isn't the same as the site setting Post_Approval so the check would return false.

I couldn't find any issues when I tested it (tested for character collisions, using a similar Post.Approval group, etc). The function is used to order the inbox and when redirecting posts to the inbox. Both seemed to work fine.